requirements.txt
flask
Flask-WTF
pandas
python-dotenv
